Anaconda rather than elpy, and simpler hydra
Didn't like the large popup box that hydra created, so using only the
'hint' version of the commands.  Most of the commands are now getting
to be muscle memory, so the HUD is not so necessary.  Moreso just
giving myself my own keymap, but with the optional fallback HUD (what
about my own map but just using guide-key?)