Remove tags from git
When using editors in both windows (WSL) and linux, i end up with a
constantly-changing tag file every time I make a commit from either
platform.

Instead, manage tags locally. Perhaps I'll modify this to be based on
git hooks, so that on checkout or commit it'll automatically generate a
local-only tags file.