Use vim-pandoc-syntax instead of plasticboy's markdown
Want to (optionally) use autowrap while in markdown, but this fails for
both tpope and plasticboy's markdown when editing lists (automatically
wraps list entries as if they are comments, e.g. join and remove '-').

Tried to keep the simpler approach, but even after playing with
'formatoptions', 'formatlistpat', and 'com' settings, couldn't get
either of those syntax plugins to properly handle lists with
autowrapping.