Vim function, and command -- 'open stack of files'
This function takes a list of files and opens then horizontally split.
If the second argument is 0, a vertical split will first be opened,
otherwise the stack will replace the currently visible splits.

Commands :Habits, :Thesis, and :Todo have been created. Habits is
daily/weekly/monthly, Thesis is general/dairy/beef, and Todo is today
and todo.txt. Using 'bang' with these (e.g. :Habits!) will  provide the
argument to the function that causes ONLY these files to be displayed.