Vim: Backlinks, gotofile, copy filename on bufleave
Backlinks will RG (ripgrep) for any file that mentions the filename of
the current file.

GotoFile will try to improve the `gf` command, as it doesn't work when
you're on the text of a markdown link, or on an earlier position on the
line. This command, and relevant keybinds, will instead find
NON-REFERENCE links. Reference links are a bit trickier to manage, so
not done yet.

BufLeave filename copying...this was a bit of an epiphany for
notetaking. This basically will make it much easier for me to link notes
together. Every time I leave a txt or markdown file, I'll automatically
copy the filename _as a markdown link_ to register a. I can then paste a
into the new file to link the notes together.

Additionally, I've added an autocommand for when a new txt or markdown
file is created, to paste what's in register a, although not sure if
I'll stick with this in practice.