markdown#goto_file edit, split, or vsplit
goto_file originally only edited the file. It was then modified to allow
'split', to make it easier when working with notes to keep the parent
but also open the 'child' for context. Now, I have the choice of split
or vsplit, to enable better readability depending on my display setup.

TODO - check if I can detect the current window size and automatically
determine when to split or vsplit. (Look at the 'fold' method which
calculates the number of visible columns before wrap).