Make 'Backlinks' look for strict string match
Previously, Backlink would add any file with a substring match to the
quickfix list (e.g. if the file was called 'abc.md', we would also
surface results that are linking to file 'blah-abc.md').

This regex looks for the opening bracket of the markdown URL, followed
by an optional './' and then the exact filename match.