Some doom-emacs config, and fix RUST_SRC_PATH
RUST_SRC_PATH, for nightly, doesn't seem to have a 'src' directory. So,
I'm instead symlinking the docs for stable to ~/.rust_src, and then
using this as the variable for RUST_SRC_PATH.

Hopefully on the next nightly release, and rustup update, this will be
sorted, as it feels a bit of a bodge.