emacs workaround for clipboard in WSL2
Basically, the socket that WSL uses to communicate between the linux VM
and the host environment is not static. Inside the terminal, things are
generally fine, but emacs launches with a different socket than the host
terminal, and it's the host terminal that receives, and acts on,
clipboard commands.

This hack basically tests all WSL interop sockets for the
launching-terminal's PID, and then exports this variable to a file. It
also echos the socket to a file, which emacs reads EVERY TIME it wants
to do a wsl-copy or wsl-paste operation.