Add shaid todo.sh plugin
Extract shaid (something like @a1b2-c3d4-e5f6-a7b8, derived from the first
16 digits of the sha256 of a project file) from a todo.

This may be useful if I want to mark a todo as being associated with a
particular sub-task, or otherwise giving it a unique identifier.

MAY be overkill...not really tried it in my workflow yet.